requirements.txt
gradio==3.24.1
sentencepiece==0.1.97
tqdm==4.65.0
llama_cpp_python==0.1.25
